fix: Delete file attachments for posts on uninstall.

+++ uninstall.php
@@ -82,12 +82,22 @@ function wpas_uninstall() {
 			$upload_dir = wp_upload_dir();
 			$dirpath    = trailingslashit( $upload_dir['basedir'] ) . "awesome-support/ticket_$post->ID";
 
-			if ( $post->post_parent == 0 ) {
+			if ( $post->post_parent == 0 && is_dir( $dirpath ) ) {
+
+				$it    = new RecursiveDirectoryIterator( $dirpath, RecursiveDirectoryIterator::SKIP_DOTS );
+				$files = new RecursiveIteratorIterator( $it, RecursiveIteratorIterator::CHILD_FIRST );
+
+				/* Delete each file */
+				foreach ( $files as $file ) {
+					if ( $file->isDir() ) {
+						rmdir( $file->getRealPath() );
+					} else {
+						unlink( $file->getRealPath() );
+					}
+				}
 
 				/* Delete the uploads folder */
-				if ( is_dir( $dirpath ) ) {
-					rmdir( $dirpath );
-				}
+				rmdir( $dirpath );
 
 				/* Remove transients */
 				delete_transient( "wpas_activity_meta_post_$post->ID" );
